RoleId Obsolete Warnings

+++ src/SFA.DAS.EmployerApprenticeshipsService.TestCommon/ScenarioCommonSteps/UserSteps.cs
@@ -76,7 +76,7 @@ namespace SFA.DAS.EAS.TestCommon.ScenarioCommonSteps
             var membershipRepository = _container.GetInstance<IMembershipRepository>();
             var userRecord = userRepository.GetUserByRef(user.UserRef).Result;
 
-            membershipRepository.Create(userRecord.Id, accountId, (short)role).Wait();
+            membershipRepository.Create(userRecord.Id, accountId, role).Wait();
 
             return userRecord.Id;
         }
